move two more functions to utils
