+++ asciidoc/asciidoc.py
@@ -21,7 +21,6 @@ import csv
 from functools import lru_cache
 import getopt
 import io
-import locale
 import os
 import re
 import shutil
@@ -1020,48 +1019,6 @@ def subs_attrs(lines, dictionary=None):
         return tuple(result)
 
 
-east_asian_widths = {
-    'W': 2,   # Wide
-    'F': 2,   # Full-width (wide)
-    'Na': 1,  # Narrow
-    'H': 1,   # Half-width (narrow)
-    'N': 1,   # Neutral (not East Asian, treated as narrow)
-    'A': 1,   # Ambiguous (s/b wide in East Asian context, narrow otherwise, but that doesn't work)
-}
-"""Mapping of result codes from `unicodedata.east_asian_width()` to character
-column widths."""
-
-
-def column_width(s):
-    width = 0
-    for c in s:
-        width += east_asian_widths[unicodedata.east_asian_width(c)]
-    return width
-
-
-def date_time_str(t):
-    """Convert seconds since the Epoch to formatted local date and time strings."""
-    source_date_epoch = os.environ.get('SOURCE_DATE_EPOCH')
-    if source_date_epoch is not None:
-        t = time.gmtime(min(t, int(source_date_epoch)))
-    else:
-        t = time.localtime(t)
-    date_str = time.strftime('%Y-%m-%d', t)
-    time_str = time.strftime('%H:%M:%S', t)
-    if source_date_epoch is not None:
-        time_str += ' UTC'
-    elif time.daylight and t.tm_isdst == 1:
-        time_str += ' ' + time.tzname[1]
-    else:
-        time_str += ' ' + time.tzname[0]
-    # Attempt to convert the localtime to the output encoding.
-    try:
-        time_str = time_str.decode(locale.getdefaultlocale()[1])
-    except Exception:
-        pass
-    return date_str, time_str
-
-
 class Lex:
     """Lexical analysis routines. Static methods and attributes only."""
     prev_element = None
@@ -1241,7 +1198,7 @@ class Document(object):
         Set implicit attributes and attributes in 'attrs'.
         """
         t = time.time()
-        self.attributes['localdate'], self.attributes['localtime'] = date_time_str(t)
+        self.attributes['localdate'], self.attributes['localtime'] = utils.date_time_str(t)
         self.attributes['asciidoc-module'] = 'asciidoc'
         self.attributes['asciidoc-version'] = VERSION
         self.attributes['asciidoc-confdir'] = CONF_DIR
@@ -1267,7 +1224,7 @@ class Document(object):
             else:
                 t = None
             if t:
-                self.attributes['docdate'], self.attributes['doctime'] = date_time_str(t)
+                self.attributes['docdate'], self.attributes['doctime'] = utils.date_time_str(t)
             if self.infile != '<stdin>':
                 self.attributes['infile'] = self.infile
                 self.attributes['indir'] = os.path.dirname(self.infile)
@@ -1919,7 +1876,7 @@ class Title:
             if len(lines) < 2:
                 return False
             title, ul = lines[:2]
-            title_len = column_width(title)
+            title_len = utils.column_width(title)
             ul_len = len(ul)
             if ul_len < 2:
                 return False
